NBA_backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Xuzhejia <Xuzhejia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
PyMySQL = "^1.1.1"
typing-extensions = "^4.12.2"
openai = "^1.59.8"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
